Comment filed test
When converting a XML to Lua and then again to XML,
the result may be a little bit different
from the original XML.

The commented test checks this.
Since it's not ensured the resulting XML in this
process will be exactly equal to the original,
the test was disabled.

This is a tricky test because
the original XML may be indented in
a different way and the resulting
XML will be perfectly idented.